python/requirements.txt
mysql-connector
requests
grafana-api
